pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "metamorphosis-backend"
version = "0.1.0"
description = "Backend service for metamorphosis project"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "python-jose>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
]
